+++ index.html
@@ -9,7 +9,7 @@
   <title>Кекстаграм</title>
 </head>
 <body>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
